% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clusters_different_shapes}
\alias{clusters_different_shapes}
\title{Generate Clusters with Different Shapes}
\usage{
clusters_different_shapes(
  sample_size,
  num_gussian_clusters,
  num_non_gaussian_clusters,
  cluster_sd_gau,
  cluster_sd_non_gau,
  num_dims,
  a,
  b
)
}
\arguments{
\item{sample_size}{The total number of data points to be generated.}

\item{num_gussian_clusters}{The number of Gaussian clusters to generate.}

\item{num_non_gaussian_clusters}{The number of non-Gaussian clusters to generate.}

\item{cluster_sd_gau}{The standard deviation for the Gaussian clusters.}

\item{cluster_sd_non_gau}{The standard deviation for the non-Gaussian clusters.}

\item{num_dims}{The number of dimensions for the data points.}

\item{a}{The scaling factor for the non-Gaussian cluster shape.}

\item{b}{The translation factor for the non-Gaussian cluster shape.}
}
\value{
A matrix containing the generated clusters with different shapes.
}
\description{
This function generates clusters with different shapes, including both Gaussian and non-Gaussian clusters.
}
\examples{
# Generate clusters with default parameters
data <- clusters_different_shapes(sample_size = 300, num_gussian_clusters = 4,
num_non_gaussian_clusters = 2, cluster_sd_gau = 0.05, cluster_sd_non_gau = 0.1,
num_dims = 7, a = 2, b = 4)
}
