% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd_obj.R
\name{conic_spiral_3d_row}
\alias{conic_spiral_3d_row}
\title{Generate points on a conic spiral in 3D space.}
\usage{
conic_spiral_3d_row(a, b, c, w)
}
\arguments{
\item{a}{Final radius of the cone.}

\item{b}{Height of the object.}

\item{c}{Inner radius.}

\item{w}{Number of spirals.}
}
\value{
A matrix containing the generated points on the conic spiral.
}
\description{
This function generates points on a conic spiral in 3D space.
}
\examples{
set.seed(20240412)
conic_spiral_3d_row(1, 2, 0.5, 3)

}
