% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{curvy_branching_with_noise}
\alias{curvy_branching_with_noise}
\title{Generate Curvy Branching Clusters with Noise}
\usage{
curvy_branching_with_noise(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{The total number of data points to be generated.}

\item{num_noise_dims}{The number of additional noise dimensions to be generated.}

\item{min_noise}{The minimum value for the noise added to the data points.}

\item{max_noise}{The maximum value for the noise added to the data points.}
}
\value{
A matrix containing the generated data, with each row representing a data point.
}
\description{
This function generates data with curvy branching clusters along with added noise.
}
\examples{

# Generate curvy branching clusters with noise with custom parameters
data <- curvy_branching_with_noise(sample_size = 200, num_noise_dims = 8,
min_noise = -0.05, max_noise = 0.05)
}
