% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{curvy_tree_with_noise}
\alias{curvy_tree_with_noise}
\title{Generate Curvy Tree Data with Noise}
\usage{
curvy_tree_with_noise(n, num_noise, min_n, max_n)
}
\arguments{
\item{n}{The total number of samples to generate.}

\item{num_noise}{The number of additional noise dimensions to add to the data.}

\item{min_n}{The minimum value for the noise dimensions.}

\item{max_n}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the curvy tree data with added noise.
}
\description{
This function generates a dataset representing a curvy tree structure, with added noise.
}
\examples{
tree_data <- curvy_tree_with_noise(n = 300, num_noise = 8,
                                   min_n = -0.05, max_n = 0.05)
}
