% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd_obj.R
\name{dini_surface_3d_row}
\alias{dini_surface_3d_row}
\title{Generate points on a Dini's surface.}
\usage{
dini_surface_3d_row(a = 1, b = 1)
}
\arguments{
\item{a}{Outer radius of the surface.}

\item{b}{Space between loops.}
}
\value{
A matrix containing the generated points on the surface.
}
\description{
This function generates points on a Dini's surface.
}
\examples{
set.seed(20240412)
dini_surface_3d_row(a = 1, b = 1)

}
