% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{eight_branching_data_with_noise}
\alias{eight_branching_data_with_noise}
\title{Generate Eight Branching Data with Noise}
\usage{
eight_branching_data_with_noise(
  sample_size,
  num_noise_dims,
  min_noise,
  max_noise
)
}
\arguments{
\item{sample_size}{The total number of samples to generate.}

\item{num_noise_dims}{The number of additional noise dimensions to add to the data.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the eight branching data with added noise.
}
\description{
This function generates a dataset representing eight branching patterns, with added noise.
}
\examples{
branching_data <- eight_branching_data_with_noise(sample_size = 400,
num_noise_dims = 8, min_noise = -0.05, max_noise = 0.05)
}
