% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{four_diff_long_clutsers_with_noise}
\alias{four_diff_long_clutsers_with_noise}
\title{Generate Four Different Long Clusters with Noise}
\usage{
four_diff_long_clutsers_with_noise(
  sample_size,
  num_noise_dims,
  min_noise,
  max_noise
)
}
\arguments{
\item{sample_size}{The total number of samples to generate.}

\item{num_noise_dims}{The number of additional noise dimensions to add to the data.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the four different long clusters with added noise.
}
\description{
This function generates a dataset consisting of four different long clusters with added noise.
}
\examples{
four_diff_long_clusters <- four_diff_long_clutsers_with_noise(sample_size = 200,
num_noise_dims = 8, min_noise = -0.05, max_noise = 0.05)
}
