% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{gaussian_clusters}
\alias{gaussian_clusters}
\title{Generate synthetic data with Gaussian clusters}
\usage{
gaussian_clusters(
  sample_size,
  with_seed = NULL,
  num_clusters,
  mean_matrix,
  var_vec,
  num_dims,
  num_noise_dims,
  min_noise,
  max_noise
)
}
\arguments{
\item{sample_size}{The number of data points to generate.}

\item{with_seed}{Optional seed value for reproducibility.}

\item{num_clusters}{The number of clusters to generate.}

\item{mean_matrix}{A matrix specifying the means for each cluster.
Each row corresponds to a cluster, and each column corresponds to a dimension.}

\item{var_vec}{A vector specifying the variance for each dimension.}

\item{num_dims}{The number of dimensions for the data.}

\item{num_noise_dims}{The number of additional noise dimensions to add.}

\item{min_noise}{The minimum value for noise generation.}

\item{max_noise}{The maximum value for noise generation.}
}
\value{
A data frame containing the synthetic data with Gaussian clusters.
}
\description{
This function generates synthetic data with Gaussian clusters.
}
\examples{

gaussian_clusters(sample_size = 300, with_seed = NULL, num_clusters = 5,
mean_matrix = rbind(c(1,0,0,0), c(0,1,0,0), c(0,0,1,0),
c(0,0,0,1), c(0,0,0,0)), var_vec = c(0.05, 0.05, 0.05, 0.05, 0.05),
num_dims = 4, num_noise_dims = 0, min_noise = -0.05, max_noise = 0.05)

}
