% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone.R
\name{gen_bluntedcorn}
\alias{gen_bluntedcorn}
\title{Generate Blunted Corn}
\usage{
gen_bluntedcorn(n = 500, p = 4, h = 5, ratio = 0.5)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{h}{A numeric value (default: 5) representing the h of the corn.}

\item{ratio}{A numeric value (default: 0.5) representing the radius tip to radius base ratio of the corn. Should be less than 1.}
}
\value{
A data containing the blunted corn.
}
\description{
This function generates a dataset representing a blunted corn.
}
\examples{
set.seed(20240412)
blunted_corn_data <- gen_bluntedcorn(n = 500, p = 4, h = 5, ratio = 0.5)
}
