% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{gen_circle}
\alias{gen_circle}
\title{Generate Circle in p-d}
\usage{
gen_circle(n = 500, p = 4)
}
\arguments{
\item{n}{An integer value (default: 500) representing the sample size.}

\item{p}{An integer value (default: 4) representing the number of dimensions.}
}
\value{
A data containing a circle.
}
\description{
This function generates a dataset representing a structure with a circle.
}
\examples{
set.seed(20240412)
circle <- gen_circle(n = 500, p = 4)
}
