% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circulars.R
\name{gen_circular_pd}
\alias{gen_circular_pd}
\title{Generate Circular in p-d}
\usage{
gen_circular_pd(
  n = 500,
  p = 4,
  shift = c(0, 0, 0),
  scale_fac = c(1, cos(0.4), sin(0.4))
)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{shift}{A numeric vector (default: c(0, 0, 0)) representing the shift along x1, x2, x3 respectively.}

\item{scale_fac}{A numeric vector (default:  c(1, cos(.4), sin(.4))) representing the scale factors along x1, x2, x3 respectively.}
}
\value{
A data containing a circular.
}
\description{
This function generates a dataset representing a structure with a circular.
}
\examples{
set.seed(20240412)
circular_data <- gen_circular_pd(n = 500, p = 4)
}
