% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gen_clustloc}
\alias{gen_clustloc}
\title{Generate Cluster Locations}
\usage{
gen_clustloc(p = 4, k = 3)
}
\arguments{
\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{k}{A numeric value (default: 3) representing the number of clusters.}
}
\value{
A matrix of the locations.
}
\description{
This function generate locations for any number of clusters in any dimensions.
}
\examples{
set.seed(20240412)
gen_clustloc(p = 4, k = 3)

}
