% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_shapes.R
\name{gen_clustRomanSurface}
\alias{gen_clustRomanSurface}
\title{Generate two roman surface shaped clusters}
\usage{
gen_clustRomanSurface(n = c(500, 500), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(500, 500)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the tow roman surface shaped clusters.
}
\description{
This function generates tow roman surface shaped clusters
}
\examples{
set.seed(20240412)
data <- gen_clustRomanSurface(n = c(500, 500), p = 4)
}
