% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular_old.R
\name{gen_clusts_tri}
\alias{gen_clusts_tri}
\title{Generate Triangular Prism clusters}
\usage{
gen_clusts_tri(n = c(200, 300, 500), p = 4, k = 3)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 500, 300)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{k}{A numeric value (default: 3) representing the number of clusters.}
}
\value{
A data containing a triangular prism clusters.
}
\description{
This function generates p-D triangular prism clusters datasets.
}
\examples{
set.seed(20240412)
triangular_clusts_data <- gen_clusts_tri(n = c(200, 300, 500), p = 4, k = 3)
}
