% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{gen_conicSpiral}
\alias{gen_conicSpiral}
\title{Generate Conical Spiral}
\usage{
gen_conicSpiral(n = 500, p = 4, spins = 1, cone_height = 2, cone_radius = 1)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{spins}{A numeric value (default: 1) representing the number of loops of the spiral.}

\item{r}{A numeric value (default: 1) representing the radius of the circle.}
}
\value{
A data containing a conical spiral structure.
}
\description{
This function generates a dataset representing a conical spiral structure.
}
\examples{
set.seed(20240412)
data <- gen_conicSpiral(n = 500, p = 4, r = 1, spins = 1)
}
