% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corn.R
\name{gen_corn_blunted}
\alias{gen_corn_blunted}
\title{Generate Blunted Corn}
\usage{
gen_corn_blunted(n = 500, p = 4, h = 5, rb = 1.5, rt = 0.8)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{h}{A numeric value (default: 5) representing the h of the corn.}

\item{rb}{A numeric value (default: 1.5) representing the base radius of the corn.}

\item{rt}{A numeric value (default: 0.8) representing the tip radius of the corn.}
}
\value{
A data containing the blunted corn.
}
\description{
This function generates a dataset representing a blunted corn.
}
\examples{
set.seed(20240412)
blunted_corn_data <- gen_corn_blunted(n = 500, p = 4, h = 5, rb = 1.5, rt = 0.8)
}
