% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corn.R
\name{gen_corn_rectangular_base}
\alias{gen_corn_rectangular_base}
\title{Generate Rectangular Based Corn}
\usage{
gen_corn_rectangular_base(n = 500, p = 4, h = 5, l_vec = c(3, 2), rt = 0.5)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{h}{A numeric value (default: 5) representing the h of the corn.}

\item{l_vec}{A numeric vector (default: c(3, 2)) representing the base lengths along the and y of the corn.}

\item{rt}{A numeric value (default: 0.5) representing the tip radius of the corn.}
}
\value{
A data containing the rectangular based corn.
}
\description{
This function generates a dataset representing a rectangular based corn.
}
\examples{
set.seed(20240412)
rectangular_corn_data <- gen_corn_rectangular_base(n = 500, p = 4, h = 5, l_vec = c(3, 2), rt = 0.5)
}
