% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corn.R
\name{gen_corn_triangular_base}
\alias{gen_corn_triangular_base}
\title{Generate Triangular Based Corn}
\usage{
gen_corn_triangular_base(n = 500, p = 4, h = 5, l = 3, rt = 0.5)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{h}{A numeric value (default: 5) representing the h of the corn.}

\item{l}{A numeric value (default: 3) representing the base length of the corn.}

\item{rt}{A numeric value (default: 0.5) representing the tip radius of the corn.}
}
\value{
A data containing the triangular based corn.
}
\description{
This function generates a dataset representing a triangular based corn.
}
\examples{
set.seed(20240412)
triangular_corn_data <- gen_corn_triangular_base(n = 500, p = 4, h = 5, l = 3, rt = 0.5)
}
