% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{gen_crescent}
\alias{gen_crescent}
\title{Generate Crescent}
\usage{
gen_crescent(n = 500)
}
\arguments{
\item{n}{An integer value (default: 500) representing the sample size.}
}
\value{
A tibble containing a Crescent structure.
}
\description{
This function generates a dataset representing a structure with a Crescent pattern.
}
\examples{
set.seed(20240412)
crescent <- gen_crescent(n = 500)
}
