% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{gen_cubehole}
\alias{gen_cubehole}
\title{Generate Cube with Hole}
\usage{
gen_cubehole(n = 500, p = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the cube data with a hole.
}
\description{
This function generates a dataset representing a cube with a hole.
}
\examples{
set.seed(20240412)
cube_hole_data <- gen_cubehole(n = 1000, p = 4)
}
