% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{gen_curve}
\alias{gen_curve}
\title{Generate Curvy}
\usage{
gen_curve(n = 500, p = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing a curvy structure.
}
\description{
This function generates a dataset representing a structure with a curvy pattern.
}
\examples{
set.seed(20240412)
curvy_data <- gen_curve(n = 500, p = 4)
}
