% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R, R/branching_old.R
\name{gen_curvy_branch_clust}
\alias{gen_curvy_branch_clust}
\title{Generate Curvy Branching Cluster Data}
\usage{
gen_curvy_branch_clust(n = c(200, 200, 100), p = 4)

gen_curvy_branch_clust(n = c(200, 200, 100), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 200, 100)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing two curvy clusters and one Gaussian cluster.

A data containing two curvy clusters and one Gaussian cluster.
}
\description{
This function generates two curvy clusters and one Gaussian cluster in the middle.

This function generates two curvy clusters and one Gaussian cluster in the middle.
}
\examples{

# Generate curvy branching cluster data with custom parameters
set.seed(20240412)
data <- gen_curvy_branch_clust(n = c(200, 200, 100), p = 4)

# Generate curvy branching cluster data with custom parameters
set.seed(20240412)
data <- gen_curvy_branch_clust(n = c(200, 200, 100), p = 4)
}
