% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{gen_curvycylinder}
\alias{gen_curvycylinder}
\title{Generate Curvy Cylinder}
\usage{
gen_curvycylinder(n = 500, h = 10)
}
\arguments{
\item{n}{An integer value (default: 500) representing the sample size.}

\item{h}{A numeric value (default: 10) representing the height of the cylinder.}
}
\value{
A data containing a curvy cylinder.
}
\description{
This function generates a dataset representing a structure with a curvy cylinder.
}
\examples{
set.seed(20240412)
curvycylinder <- gen_curvycylinder(n = 500, h = 10)
}
