% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{gen_eightbranches}
\alias{gen_eightbranches}
\title{Generate data with eight branches}
\usage{
gen_eightbranches(n = c(200, 300, 150, 250, 100, 100, 100, 100), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 300, 150, 250, 100, 100, 100, 100)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing eight branches.
}
\description{
This function generates a dataset representing a structure with eight branches.
}
\examples{
set.seed(20240412)
branching_data <- gen_eightbranches(n = c(200, 300, 150, 250, 100, 100, 100, 100), p = 4)
}
