% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corn.R
\name{gen_filled_hexagonal_pyramid}
\alias{gen_filled_hexagonal_pyramid}
\title{Generate Hexagonal Based Corn}
\usage{
gen_filled_hexagonal_pyramid(n = 500, p = 4, h = 5, rb = 3)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{h}{A numeric value (default: 5) representing the h of the corn.}

\item{rb}{A numeric value (default: 3) representing the base radius of the corn.}
}
\value{
A data containing the hexagonal based corn.
}
\description{
This function generates a dataset representing a hexagonal based corn.
}
\examples{
set.seed(20240412)
hexagonal_corn_data <- gen_filled_hexagonal_pyramid(n = 500, p = 4, h = 5, rb = 3)
}
