% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{gen_fivebranches}
\alias{gen_fivebranches}
\title{Generate data with five branches}
\usage{
gen_fivebranches(n = c(200, 100, 300, 400, 300), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 100, 300, 400, 300)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing five branches.
}
\description{
This function generates a dataset representing a structure with five branches.
}
\examples{
set.seed(20240412)
tree_data <- gen_fivebranches(n = c(200, 100, 300, 400, 300), p = 4)
}
