% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_shapes.R
\name{gen_four_long_clusts}
\alias{gen_four_long_clusts}
\title{Generate Four Different Long Clusters}
\usage{
gen_four_long_clusts(n = c(200, 150, 300, 150), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 150, 300, 150)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the four different long clusters.
}
\description{
This function generates a dataset consisting of four different long clusters.
}
\examples{
set.seed(20240412)
four_diff_long_clusters <- gen_four_long_clusts(n = c(200, 150, 300, 150), p = 4)
}
