% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{gen_fourbranches}
\alias{gen_fourbranches}
\title{Generate data with four branches}
\usage{
gen_fourbranches(n = c(200, 300, 150, 250), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 300, 150, 250)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing four branches.
}
\description{
This function generates a dataset representing a structure with four branches.
}
\examples{
set.seed(20240412)
four_branching_data <- gen_fourbranches(n = c(200, 300, 150, 250), p = 4)
}
