% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gen_globalminmax}
\alias{gen_globalminmax}
\title{Generate Global Minimum and Maximum}
\usage{
gen_globalminmax(data_list, p)
}
\arguments{
\item{data_list}{A matrix list containing data structures.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A list containing global minimum and maximum.
}
\description{
This function generates a list that contains global minimum and maximum.
}
\examples{
set.seed(20240412)
data1 <- gen_gaussian(n= 500, p = 4)
data2 <- gen_bluntedcorn(n= 300, p = 4)
dfs <- list(data1, data2)
gen_globalminmax(data_list = dfs, p = 4)

}
