% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{gen_gridedSphere}
\alias{gen_gridedSphere}
\title{Generate Grided Sphere}
\usage{
gen_gridedSphere(n = 500, p = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing a grided sphere.
}
\description{
This function generates a dataset representing a structure with a grided sphere.
}
\examples{
set.seed(20240412)
sphere_data <- gen_gridedSphere(n = 500, p = 4)
head(sphere_data, 5)
}
