% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{gen_helicalHyperspiral}
\alias{gen_helicalHyperspiral}
\title{Generate Helical Hyper Spiral}
\usage{
gen_helicalHyperspiral(n = 500, p = 4, a = 0.05, b = 0.1, c = 1, r = 0.5)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{a}{A numeric value (default: 0.05) representing the scale factor along x2.}

\item{b}{A numeric value (default: 0.1) representing the scale factor along x3.}

\item{c}{A numeric value (default: 1) representing the scale factor of sine along x3.}

\item{r}{A numeric value (default: 0.5) representing the radius of the spiral.}
}
\value{
A data containing a helical hyper spiral.
}
\description{
This function generates a dataset representing a structure with a helical hyper spiral.
}
\examples{
set.seed(20240412)
data <- gen_helicalHyperspiral(n = 500, p = 4, a = 0.05, b = 0.1, c = 1, r = 0.5)
}
