% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{gen_helicalspiral}
\alias{gen_helicalspiral}
\title{Generate Helical Hyper Spiral}
\usage{
gen_helicalspiral(n = 500)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}
}
\value{
A data containing a helical hyper spiral.
}
\description{
This function generates a dataset representing a structure with a helical hyper spiral.
}
\examples{
set.seed(20240412)
helicalspiral <- gen_helicalspiral(n = 500)
}
