% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{gen_hollowsphere}
\alias{gen_hollowsphere}
\title{Generate Hollow Sphere}
\usage{
gen_hollowsphere(n = 500, p = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing a hollow sphere.
}
\description{
This function generates a dataset representing a structure with a sphere with points on the surface.
}
\examples{
set.seed(20240412)
hollowsphere <- gen_hollowsphere(n = 500)
}
