% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{gen_hyperplane}
\alias{gen_hyperplane}
\title{Generate points on a plane in high-dimensions}
\usage{
gen_hyperplane(n = 500, p = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the generated points on the plane.
}
\description{
This function generates points on a plane in high-dimensions.
}
\examples{
set.seed(20240412)
plane_points <- gen_hyperplane(n = 500, p = 4)

}
