% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{gen_hyperplaneHole}
\alias{gen_hyperplaneHole}
\title{Generate Hyper Plane with Hole}
\usage{
gen_hyperplaneHole(n = 500, p = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the hyperplane data with a hole.
}
\description{
This function generates a dataset representing a hyper plane with a hole.
}
\examples{
set.seed(20240412)
plane_hole_data <- gen_hyperplaneHole(n = 1000, p = 4)
}
