% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd_obj.R
\name{gen_intersect_3d_two_dini_surface}
\alias{gen_intersect_3d_two_dini_surface}
\title{Generate intersected tow dini surface shapes}
\usage{
gen_intersect_3d_two_dini_surface(n = c(500, 300), p = 3)
}
\arguments{
\item{n}{A numeric vector (default: c(500, 300)) representing the sample sizes.}

\item{p}{A numeric value (default: 3) representing the number of dimensions.}
}
\value{
A data containing the two intersected dini surface shapes.
}
\description{
This function generates intersected tow dini surface shapes.
}
\examples{
set.seed(20240412)
gen_intersect_3d_two_dini_surface(n = c(500, 300), p = 3)
}
