% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circulars.R
\name{gen_linked_clusts_circulars}
\alias{gen_linked_clusts_circulars}
\title{Generate Linked Any number of 3-D Circle Clusters}
\usage{
gen_linked_clusts_circulars(n = c(200, 500, 300), p = 4, k = 3)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 500, 300)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{k}{A numeric value (default: 3) representing the number of clusters.}
}
\value{
A data containing the any number of overlapped circle clusters.
}
\description{
This function generates a dataset representing a structure with any number of overlapped circles.
}
\examples{

# Generate linked data with noise with custom parameters
set.seed(20240412)
data <- gen_linked_clusts_circulars(n = c(200, 500, 300), p = 4, k = 3)
}
