% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_shapes.R
\name{gen_longClust}
\alias{gen_longClust}
\title{Generate Long Cluster Data}
\usage{
gen_longClust(n = c(200, 500, 300), p = 4, k = 3)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 500, 300)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{k}{A numeric value (default: 3) representing the number of clusters.}
}
\value{
A data containing the long cluster data.
}
\description{
This function generates a dataset consisting of any number of long clusters.
}
\examples{
set.seed(20240412)
long_cluster <- gen_longClust(n = c(200, 500, 300), p = 4, k = 3)
}
