% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{gen_longlinear}
\alias{gen_longlinear}
\title{Generate Long Linear Data}
\usage{
gen_longlinear(n = 500, p = 4)
}
\arguments{
\item{n}{An integer value (default: 500) representing the sample size.}

\item{p}{An integer value (default: 4) representing the number of dimensions.}
}
\value{
A tibble containing the long linear data.
}
\description{
This function generates a dataset consisting of long linear data.
}
\examples{
set.seed(20240412)
longlinear <- gen_longlinear(n = 500, p = 4)
}
