% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobius.R
\name{gen_mobius}
\alias{gen_mobius}
\title{Generate a 3-D Mobius}
\usage{
gen_mobius(n = 500)
}
\arguments{
\item{n}{An integer value (default: 500) representing the sample size.}
}
\value{
A tibble containing a mobius structure.
}
\description{
This function generates a dataset representing a structure with a mobius.
}
\examples{
set.seed(20240412)
mobius <- gen_mobius(n = 500)
}
