% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{gen_one_grid}
\alias{gen_one_grid}
\title{Generate One Grid Data}
\usage{
gen_one_grid(n = c(10, 10, 10, 10), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(10, 10)) representing the number of grid points along each axes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the one grid data.
}
\description{
This function generates a grid dataset with specified grid points along each axes..
}
\examples{
set.seed(20240412)
one_grid <- gen_one_grid(n = c(10, 10, 10, 10), p = 4)
}
