% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{gen_one_grid_with_bkg}
\alias{gen_one_grid_with_bkg}
\title{Generate One Grid with Background Noise}
\usage{
gen_one_grid_with_bkg(n = c(10, 10, 10, 10), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(10, 10)) representing the number of grid points along each axes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing one grid data with background noise.
}
\description{
This function generates a grid data and background noise.
}
\examples{
set.seed(20240412)
one_grid_bkg <- gen_one_grid_with_bkg(n = c(10, 10, 10, 10), p = 4)
}
