% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{gen_orgcurvybranches}
\alias{gen_orgcurvybranches}
\title{Generate data with curvy shaped branches in a initial point}
\usage{
gen_orgcurvybranches(n = 400, p = 4, k = 4, allow_share = TRUE)
}
\arguments{
\item{n}{A numeric value (default: 400) representing the sample size.}

\item{k}{A numeric value (default: 4) representing the number of branches.}

\item{allow_share}{A logical value (default: TRUE). If TRUE, multiple branches may share the same 2D subspace. If FALSE, branches are sampled without replacement from all possible 2D subspaces until exhausted.}
}
\value{
A data containing curvy shaped branches originated in one point.
}
\description{
This function generates a dataset representing a structure with curvy shaped branches.
}
\examples{
set.seed(20240412)
orgcurvybranches <- gen_orgcurvybranches(n = 400, k = 4)
}
