% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_shapes.R
\name{gen_overlapConicSpirals}
\alias{gen_overlapConicSpirals}
\title{Generate overlapped two conic spiral shapes}
\usage{
gen_overlapConicSpirals(n = c(500, 500), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(500, 500)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing a two overlapped conic spiral shapes.
}
\description{
This function generates overlapped tow conic spiral shapes.
}
\examples{
set.seed(20240412)
data <- gen_overlapConicSpirals(n = c(500, 500), p = 4)

}
