% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_shapes.R
\name{gen_overlapDiniSurface}
\alias{gen_overlapDiniSurface}
\title{Generate intersected two dini surface shapes}
\usage{
gen_overlapDiniSurface(n = c(500, 500), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(500, 500)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the two intersected dini surface shapes.
}
\description{
This function generates intersected tow dini surface shapes.
}
\examples{
set.seed(20240412)
data <- gen_overlapDiniSurface(n = c(500, 500), p = 4)
}
