% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_shapes.R
\name{gen_overlappedCurvyCycle}
\alias{gen_overlappedCurvyCycle}
\title{Generate Overlapped Any number of Curvy Circle Clusters}
\usage{
gen_overlappedCurvyCycle(n = c(200, 500, 300), p = 4, k = 3)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 500, 300)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{k}{A numeric value (default: 3) representing the number of clusters.}
}
\value{
A data containing the any number of overlapped curvy circle clusters.
}
\description{
This function generates a dataset representing a structure with any number of overlapped curvy circles.
}
\examples{
set.seed(20240412)
curvy_cell_cycle_data <- gen_overlappedCurvyCycle(n = c(200, 500, 300), p = 4, k = 3)
}
