% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_shapes.R
\name{gen_overlapTorus}
\alias{gen_overlapTorus}
\title{Generate overlapped two torus shapes}
\usage{
gen_overlapTorus(n = c(500, 500), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(500, 500)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the two overlapped torus shapes.
}
\description{
This function generates overlapped tow torus shapes.
}
\examples{
set.seed(20240412)
data <- gen_overlapTorus(n = c(500, 500), p = 4)

}
