% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{gen_pyr}
\alias{gen_pyr}
\title{Generate p-D Triangular Pyramid}
\usage{
gen_pyr(n = 500, p = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing a triangular pyramid in p-D.
}
\description{
This function generates p-D triangular pyramid datasets.
}
\examples{
set.seed(20240412)
triangular_data <- gen_pyr(n = 500, p = 4)
}
