% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{gen_pyrHoles}
\alias{gen_pyrHoles}
\title{Generate p-D Triangular Pyramid With Triangular Pyramid shaped holes}
\usage{
gen_pyrHoles(n = 500, p = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing a triangular pyramid with triangular pyramid shaped holes.
}
\description{
This function generates p-D triangular pyramid with triangular pyramid shaped holes.
}
\examples{
set.seed(20240412)
triangular_data <- gen_pyrHoles(n = 500, p = 3)
}
