% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{gen_quadratic}
\alias{gen_quadratic}
\title{Generate Quadratic}
\usage{
gen_quadratic(n = 500, range = c(-1, 1))
}
\arguments{
\item{n}{A integer value (default: 500) representing the sample size.}

\item{range}{A numeric vector (default: c(-1, 1)) representing the range along x1 axis.}
}
\value{
A tibble containing a quadratic structure.
}
\description{
This function generates a dataset representing a structure with a quadratic pattern.
}
\examples{
set.seed(20240412)
quadratic <- gen_quadratic(n = 500)
}
