% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd_obj.R
\name{gen_roman_surface_3d_clusts}
\alias{gen_roman_surface_3d_clusts}
\title{Generate tow roman surface shaped clusters}
\usage{
gen_roman_surface_3d_clusts(n = c(500, 300), p = 3)
}
\arguments{
\item{n}{A numeric vector (default: c(500, 300)) representing the sample sizes.}

\item{p}{A numeric value (default: 3) representing the number of dimensions.}
}
\value{
A data containing the tow roman surface shaped clusters.
}
\description{
This function generates tow roman surface shaped clusters
}
\examples{
set.seed(20240412)
gen_roman_surface_3d_clusts(n = c(500, 300), p = 3)
}
