% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scurve.R
\name{gen_scurve}
\alias{gen_scurve}
\title{Generate S-curve Data}
\usage{
gen_scurve(n = 500)
}
\arguments{
\item{n}{An integer value (default: 500) representing the sample size.}
}
\value{
A data containing the generated S-curve data.
}
\description{
This function generates S-curve data.
}
\examples{
set.seed(20240412)
scurve <- gen_scurve(n = 500)
}
\references{
Buitinck, L., Louppe, G., Blondel, M., Pedregosa, F., Mueller, A.,
Grisel, O., ... & Varoquaux, G. (2013).
API design for machine learning software: experiences from the scikit-learn
project.
\emph{arXiv preprint} \emph{arXiv:1309.0238}.
}
