% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{gen_sevenbranches}
\alias{gen_sevenbranches}
\title{Generate data with seven branches}
\usage{
gen_sevenbranches(n = c(200, 100, 250, 300, 150, 400, 50), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 100, 250, 300, 150, 400, 50)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing seven branches.
}
\description{
This function generates a dataset representing a structure with seven branches.
}
\examples{
set.seed(20240412)
seven_branching_data <- gen_sevenbranches(
n = c(200, 100, 250, 300, 150, 400, 50), p = 4)
}
