% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{gen_sphericalSpiral}
\alias{gen_sphericalSpiral}
\title{Generate Spherical Spiral}
\usage{
gen_sphericalSpiral(n = 500, p = 4, r = 1, spins = 1)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{r}{A numeric value (default: 1) representing the radius of the circle.}

\item{spins}{A numeric value (default: 1) representing the number of loops of the spiral.}
}
\value{
A data containing a spherical spiral.
}
\description{
This function generates a dataset representing a structure with a spherical spiral.
}
\examples{
set.seed(20240412)
data <- gen_sphericalSpiral(n = 500, p = 4, r = 1, spins = 1)
}
