% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_shapes.R
\name{gen_three_angled_long_clusts}
\alias{gen_three_angled_long_clusts}
\title{Generate Three Different Linear Data}
\usage{
gen_three_angled_long_clusts(n = c(200, 300, 150), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 300, 150)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the three different linear data.
}
\description{
This function generates a dataset consisting of three different linear clusters.
}
\examples{
set.seed(20240412)
three_diff_linear <- gen_three_angled_long_clusts(n = c(200, 300, 150), p = 4)
}
