% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_clusters.R
\name{gen_three_clust_19}
\alias{gen_three_clust_19}
\title{Generate Three Different Shaped Clusters}
\usage{
gen_three_clust_19(n = c(700, 300, 500))
}
\arguments{
\item{n}{A numeric vector (default: c(700, 300, 500)) representing the sample sizes.}
}
\value{
A data containing the three different shaped clusters.
}
\description{
This function generates three different shaped clusters.
}
\examples{
set.seed(20240412)
gen_three_clust_19(n = c(700, 300, 500))
}
