% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{gen_threebranches}
\alias{gen_threebranches}
\title{Generate data with three branches}
\usage{
gen_threebranches(n = c(200, 500, 300), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 500, 300)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing three branches.
}
\description{
This function generates a dataset representing a structure with three branches.
}
\examples{
set.seed(20240412)
tree_data <- gen_threebranches(n = c(200, 500, 300), p = 4)
}
