% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trefoil.R
\name{gen_trefoil4d}
\alias{gen_trefoil4d}
\title{Generate 4-D Trefoil Knot Coordinates}
\usage{
gen_trefoil4d(n = 500, steps = 5)
}
\arguments{
\item{n}{An integer value (default: 500) representing the sample size.}

\item{steps}{A numeric value (default: 5) representing the number of steps for the theta parameter.}
}
\value{
A tibble containing 4-D trefoil knot.
}
\description{
This function generates coordinates for a 4-D trefoil knot.
The number of points is determined by the length of the theta and phi sequences.
}
\examples{
set.seed(20240412)
trefoil4d <- gen_trefoil4d(n = 500, steps = 5)
}
