% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular_old.R
\name{gen_tri_prism}
\alias{gen_tri_prism}
\title{Generate p-D Triangular Prism}
\usage{
gen_tri_prism(n = 500, p = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing a triangular prism in p-D.
}
\description{
This function generates p-D triangular prism datasets.
}
\examples{
set.seed(20240412)
triangular_data <- gen_tri_prism(n = 300, p = 3)
}
