% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{gen_two_curvy}
\alias{gen_two_curvy}
\title{Generate Curvy Branching Clusters with Noise}
\usage{
gen_two_curvy(n = c(300, 200), p = 4)
}
\arguments{
\item{n}{The total number of data points to be generated.}

\item{num_noise}{The number of additional noise dimensions to be generated.}

\item{min_n}{The minimum value for the noise added to the data points.}

\item{max_n}{The maximum value for the noise added to the data points.}
}
\value{
A matrix containing the generated data, with each row representing a data point.
}
\description{
This function generates data with curvy branching clusters along with added noise.
}
\examples{

# Generate curvy branching clusters with noise with custom parameters
set.seed(20240412)
data <- gen_two_curvy(n = c(300, 200), p = 4)
}
