% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{gen_two_grids_comb_bkg}
\alias{gen_two_grids_comb_bkg}
\title{Generate Two Grids with Background Noise}
\usage{
gen_two_grids_comb_bkg(n = c(10, 10, 10, 10), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(10, 10)) representing the number of grid points along each axes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing two grids data with background noise.
}
\description{
This function generates two grids with background noise.
}
\examples{
set.seed(20240412)
two_grid_comb_bkg <- gen_two_grids_comb_bkg(n = c(10, 10, 10, 10), p = 4)
}
