% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_shapes.R
\name{gen_two_long_clusts}
\alias{gen_two_long_clusts}
\title{Generate Long Cluster Data}
\usage{
gen_two_long_clusts(n = c(200, 300), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 300)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the long cluster data.
}
\description{
This function generates a dataset consisting of two long clusters.
}
\examples{
set.seed(20240412)
long_cluster <- gen_two_long_clusts(n = c(200, 300), p = 4)
}
