% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{gen_unifcube}
\alias{gen_unifcube}
\title{Generate Cube with uniform points}
\usage{
gen_unifcube(n = 625, p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(10, 10)) representing the number of grid points along each axes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the cube with uniform points.
}
\description{
This function generates a grid dataset with specified uniform points along each axes..
}
\examples{
set.seed(20240412)
unif_cube <- gen_unifcube(n = 625, p = 4)
}
