% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{gen_unifcubehole}
\alias{gen_unifcubehole}
\title{Generate Cube with Hole}
\usage{
gen_unifcubehole(n = 5000, p = 4, r_hole = 0.5)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{r_hole}{A numeric value (default: 0.5) representing the radius of the hole.}
}
\value{
A data containing the cube data with a hole.
}
\description{
This function generates a dataset representing a cube with a hole.
}
\examples{
set.seed(20240412)
cubehole <- gen_cubehole(n = 1000, p = 4)
}
