% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gen_wavydims1}
\alias{gen_wavydims1}
\title{Generate Random Noise Dimensions With Wavy Pattern}
\usage{
gen_wavydims1(n = 500, p = 4, theta = seq(pi/6, 12 * pi/6, length.out = 500))
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{theta}{A numeric vector representing the nonlinearity along each dimensions.}
}
\value{
A data containing the generated random noise dimensions.
}
\description{
This function generates random noise dimensions by adding wavy patterns.
}
\examples{
set.seed(20240412)
gen_wavydims1(n = 500, p = 4, theta = seq(pi / 6, 12 * pi / 6, length.out = 500))

}
