% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circulars.R
\name{link_data}
\alias{link_data}
\title{Generate Linked Data}
\usage{
link_data(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{The total number of data points to be generated. Should be a product of two.}

\item{num_noise_dims}{The number of additional noise dimensions to be generated.}

\item{min_noise}{The minimum value for the noise added to the data points.}

\item{max_noise}{The maximum value for the noise added to the data points.}
}
\value{
A matrix containing the generated linked data points.
}
\description{
This function generates linked data points.
}
\examples{

# Generate linked data with noise with custom parameters
data <- link_data(sample_size = 200, num_noise_dims = 4,
min_noise = -0.05, max_noise = 0.05)
}
