% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{long_cluster_data}
\alias{long_cluster_data}
\title{Generate Long Cluster Data}
\usage{
long_cluster_data(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{The total number of samples to generate.}

\item{num_noise_dims}{The number of additional noise dimensions to add to the data.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the long cluster data with added noise.
}
\description{
This function generates a dataset consisting of two long clusters with added noise.
}
\examples{
long_cluster <- long_cluster_data(sample_size = 200, num_noise_dims = 8,
                                  min_noise = -0.05, max_noise = 0.05)
}
