% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_shapes.R
\name{make_mobiusgau}
\alias{make_mobiusgau}
\title{Generate Gaussian cluster with the Mobius Cluster}
\usage{
make_mobiusgau(n = c(200, 100), p = 4)
}
\arguments{
\item{n}{An integer vector (default: c(200, 100)) representing the sample sizes.}

\item{p}{An integer value (default: 4) representing the number of dimensions.}
}
\value{
A tibble containing the mobius cluster and Gaussian cluster.
}
\description{
This function generates a dataset consisting of a mobius cluster and Gaussian cluster.
}
\examples{
mobgau <- make_mobiusgau(n = c(200, 100), p = 4)
}
