% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_clusters.R
\name{make_three_clust_16}
\alias{make_three_clust_16}
\title{Generate a Three-Cluster Dataset in High Dimensions}
\usage{
make_three_clust_16(n = c(700, 300, 500), p = 4)
}
\arguments{
\item{n}{An integer vector of length 3 specifying the number of points in
each cluster. Default is \code{c(700, 300, 500)}.}

\item{p}{Integer, the dimensionality of the space. Default is \code{4}.}
}
\value{
A data frame (or tibble, depending on `gen_multicluster()`)
  containing the generated dataset with cluster assignments.
}
\description{
This function generates a dataset consisting of three distinct clusters
in a \eqn{p}-dimensional space. Each cluster is generated with a specified
shape and location using the underlying `gen_multicluster()` function.
}
\examples{
# Generate default three clusters in 4-D
three_clust_16 <- make_three_clust_16()

}
