% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_curve.R
\name{mirror_s_curves_with_noise}
\alias{mirror_s_curves_with_noise}
\title{Generate Mirror S-curve Datasets with Noise}
\usage{
mirror_s_curves_with_noise(n, num_noise, min_n, max_n)
}
\arguments{
\item{n}{The total number of samples to generate (should be divisible by 2).}

\item{num_noise}{The number of additional noise dimensions to add to the data.}

\item{min_n}{The minimum value for the noise dimensions.}

\item{max_n}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the combined mirror S-curve datasets with added noise.
}
\description{
This function generates mirror S-curve datasets with added noise dimensions.
}
\examples{
mirror_s_curve_data <- mirror_s_curves_with_noise(n = 200, num_noise = 8,
min_n = -0.5, max_n = 0.5)
}
