% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobius_clust.R
\name{mobius_clust}
\alias{mobius_clust}
\title{Generate Mobius Cluster with Noise}
\usage{
mobius_clust(n, num_noise, min_n, max_n)
}
\arguments{
\item{n}{The total number of samples to generate.}

\item{num_noise}{The number of additional noise dimensions to add to the data.}

\item{min_n}{The minimum value for the noise dimensions.}

\item{max_n}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the mobius cluster with added noise.
}
\description{
This function generates a dataset consisting of a mobius cluster with added noise.
}
\examples{
mobius_cluster <- mobius_clust(
  n = 200, num_noise = 2, min_n = -0.05,
  max_n = 0.05
)
}
