% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobius_clust.R
\name{mobius_gau_clust}
\alias{mobius_gau_clust}
\title{Generate Gaussian cluster with the Mobius Cluster}
\usage{
mobius_gau_clust(n = c(200, 100), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 100)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the mobius cluster and Gaussian cluster.
}
\description{
This function generates a dataset consisting of a mobius cluster and Gaussian cluster.
}
\examples{
mobius_cluster <- mobius_gau_clust(n = c(200, 100), p = 4)
}
