% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mobiusgau_tsne3}
\alias{mobiusgau_tsne3}
\title{tSNE embedding for mobiusgau dataset which with noise dimensions
tSNE parameters set to perplexity: 5.}
\format{
## `mobiusgau_tsne3`
A data frame with 1000 rows and 4 columns:
\describe{
  \item{emb1}{Numeric, first tSNE 2D embeddings.}
  \item{emb2}{Numeric, second tSNE 2D embeddings.}
}
}
\source{
This dataset is generated for illustrative purposes.
}
\usage{
data(mobiusgau_tsne3)
}
\description{
The `mobiusgau_tsne3` dataset contains the tSNE (t-distributed Stochastic
Neighbor Embedding) embeddings of a five-dimensional mobiusgau.
Each data point is represented by two tSNE coordinates (emb1
and emb2).
}
\examples{
# Load the mobiusgau_tsne1 dataset
data(mobiusgau_tsne3)

# Display the first few rows of the dataset
head(mobiusgau_tsne3)

}
\keyword{datasets}
