% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mobiusgau_umap1}
\alias{mobiusgau_umap1}
\title{UMAP embedding for mobiusgau dataset which with noise dimensions
UMAP parameters set to n-neigbors: 15 and min-dist: 0.1.}
\format{
## `mobiusgau_umap1`
A data frame with 1000 rows and 4 columns:
\describe{
  \item{emb1}{Numeric, first UMAP 2D embeddings.}
  \item{emb2}{Numeric, second UMAP 2D embeddings.}
}
}
\source{
This dataset is generated for illustrative purposes.
}
\usage{
data(mobiusgau_umap1)
}
\description{
The `mobiusgau_umap1` dataset contains the UMAP (Uniform Manifold
Approximation and Projection) embeddings of a five-dimensional mobiusgau.
Each data point is represented by two UMAP coordinates (emb1
and emb2).
}
\examples{
# Load the mobiusgau_umap1 dataset
data(mobiusgau_umap1)

# Display the first few rows of the dataset
head(mobiusgau_umap1)

}
\keyword{datasets}
