% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear.R
\name{nonlinear_2d}
\alias{nonlinear_2d}
\title{Generate points on a nonlinear 2D manifold}
\usage{
nonlinear_2d(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{The number of points to generate.}

\item{num_noise_dims}{The number of noise dimensions to add to the generated points.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the generated points on the nonlinear 2D manifold.
}
\description{
This function generates points on a nonlinear 2D manifold based on a given equation.
}
\examples{
nonlinear_points <- nonlinear_2d(sample_size = 100, num_noise_dims = 2,
                                 min_noise = -1, max_noise = 1)

}
