% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{one_doublet_dfifferent_var_clusters_with_noise}
\alias{one_doublet_dfifferent_var_clusters_with_noise}
\title{Generate Doublets with Different Variance Clusters and Noise}
\usage{
one_doublet_dfifferent_var_clusters_with_noise(
  sample_size,
  num_noise_dims,
  min_noise,
  max_noise
)
}
\arguments{
\item{sample_size}{The total number of data points to be generated.}

\item{num_noise_dims}{The number of additional noise dimensions to be generated.}

\item{min_noise}{The minimum value for the noise added to the data points.}

\item{max_noise}{The maximum value for the noise added to the data points.}
}
\value{
A matrix containing the generated data, with each row representing a data point.
}
\description{
This function generates data with one set of doublets (pairs of clusters) having clusters with different variance, along with added background noise.
}
\examples{

# Generate doublets with different variance clusters and noise with custom parameters
data <- one_doublet_dfifferent_var_clusters_with_noise(sample_size = 260, num_noise_dims = 2,
min_noise = -0.05, max_noise = 0.05)
}
