% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{one_doublet_four_clusts}
\alias{one_doublet_four_clusts}
\title{Generate Doublets with Four Clusters and Noise}
\usage{
one_doublet_four_clusts(n, num_noise, min_n, max_n)
}
\arguments{
\item{n}{The total number of data points to be generated.}

\item{num_noise}{The number of additional noise dimensions to be generated.}

\item{min_n}{The minimum value for the noise added to the data points.}

\item{max_n}{The maximum value for the noise added to the data points.}
}
\value{
A matrix containing the generated data, with each row representing a data point.
}
\description{
This function generates data with one set of doublets (pairs of clusters) containing
four clusters, along with added background noise.
}
\examples{

# Generate doublets with four clusters and noise with custom parameters
set.seed(20240412)
data <- one_doublet_four_clusts(
  n = 440, num_noise = 2,
  min_n = -0.05, max_n = 0.05
)
}
