% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{one_grid_diff}
\alias{one_grid_diff}
\title{Generate One Grid with Different Offset}
\usage{
one_grid_diff(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{The total number of samples in the dataset.}

\item{num_noise_dims}{The number of additional noise dimensions to add to the data.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A list containing the grid dataset with different offsets and the sample size.
}
\description{
This function generates a single grid dataset with a different offset.
}
\examples{
one_grid_diff <- one_grid_diff(sample_size = 200, num_noise_dims = 2,
min_noise = -0.05, max_noise = 0.05)
}
