% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{one_grid_diff_with_bkg_noise}
\alias{one_grid_diff_with_bkg_noise}
\title{Generate One Grid with Different Values and Background Noise}
\usage{
one_grid_diff_with_bkg_noise(
  n = 260,
  num_noise = 5,
  min_n = -0.05,
  max_n = 0.05
)
}
\arguments{
\item{n}{The total number of samples, including the background noise.}

\item{num_noise}{The number of additional noise dimensions to add to the data.}

\item{min_n}{The minimum value for the noise dimensions.}

\item{max_n}{The maximum value for the noise dimensions.}
}
\value{
A list containing the grid dataset with different values and background noise.
}
\description{
This function generates a grid dataset with different values and background noise.
}
\examples{
one_grid_diff_with_bkg_noise <- one_grid_diff_with_bkg_noise(n = 260,
num_noise = 5, min_n = -0.05, max_n = 0.05)
}
