% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{plane_2d_with_hole}
\alias{plane_2d_with_hole}
\title{Generate 2D Plane with Hole and Noise}
\usage{
plane_2d_with_hole(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{The total number of samples to generate.}

\item{num_noise_dims}{The number of additional noise dimensions to add to the data.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A list containing the 2D plane data with a hole and the sample size.
}
\description{
This function generates a dataset representing a 2D plane with a hole in the middle, with added noise.
}
\examples{
plane_data <- plane_2d_with_hole(sample_size = 100, num_noise_dims = 2,
min_noise = 0, max_noise = 1)
}
