% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{randomize_rows}
\alias{randomize_rows}
\title{Randomize Rows of a Data Frame}
\usage{
randomize_rows(data)
}
\arguments{
\item{data}{A data frame to be randomized.}
}
\value{
A data frame with rows randomly shuffled.
}
\description{
This function randomly shuffles the rows of a given data frame.
}
\examples{
randomize_rows(mobius_clust_data)
}
