% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd_obj.R
\name{roman_surface_3d_row}
\alias{roman_surface_3d_row}
\title{Generate points on a Roman surface in 3D space.}
\usage{
roman_surface_3d_row(a = 1)
}
\arguments{
\item{a}{Maximum radius of the object.}
}
\value{
A matrix containing the generated points on the Roman surface in 3D space.
}
\description{
This function generates points on a Roman surface in 3D space.
}
\examples{
set.seed(20240412)
roman_surface_3d_row(a = 1)

}
