% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd_obj.R
\name{roman_surface_with_noise}
\alias{roman_surface_with_noise}
\title{Generate data points on a Roman surface with optional noise.}
\usage{
roman_surface_with_noise(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{Total number of data points to generate.}

\item{num_noise_dims}{Number of additional noise dimensions to add to the data.}

\item{min_noise}{Minimum value for the noise added to the data.}

\item{max_noise}{Maximum value for the noise added to the data.}
}
\value{
A matrix containing the generated data points with or without added noise.
}
\description{
This function generates data points on a Roman surface with optional noise.
}
\examples{
roman_surface_with_noise(sample_size = 100, num_noise_dims = 8,
min_noise = -0.05, max_noise = 0.05)

}
