% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_curve.R
\name{s_curve}
\alias{s_curve}
\title{Generate S-curve Data}
\usage{
s_curve(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{The number of samples to generate.}

\item{num_noise_dims}{The number of additional noise dimensions to add to the data.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the generated S-curve data.
}
\description{
This function generates S-curve data, which is a commonly used dataset for
testing and visualizing dimensionality reduction algorithms.
}
\examples{
s_curve_data <- s_curve(sample_size = 100, num_noise_dims = 3,
min_noise = -0.5, max_noise = 0.5)
}
