% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{scale_data}
\alias{scale_data}
\title{Generate Scaled data}
\usage{
scale_data(data, weights, global_val)
}
\arguments{
\item{data}{A matrix representing the data which needed to be scaled.}

\item{weights}{A numeric vector containing weights for each dimensions.}

\item{global_val}{A numeric list containing global minimum and maximum.}
}
\value{
A scaled data.
}
\description{
This function generates a scaled data.
}
\examples{
set.seed(20240412)
data1 <- gen_gaussian(n= 500, p = 4)
data2 <- gen_bluntedcorn(n= 300, p = 4)
dfs <- list(data1, data2)
global_minmax <- gen_globalminmax(data_list = dfs, p = 4)
scale_data(data = data2, weights = c(1.0, 1.0, 1.0, 0.5),
global_val = global_minmax)

}
