% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{seven_branching_data_with_noise}
\alias{seven_branching_data_with_noise}
\title{Generate Seven-Branching Data with Noise}
\usage{
seven_branching_data_with_noise(n, num_noise, min_n, max_n)
}
\arguments{
\item{n}{The total number of samples to generate.}

\item{num_noise}{The number of additional noise dimensions to add to the data.}

\item{min_n}{The minimum value for the noise dimensions.}

\item{max_n}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the seven-branching data with added noise.
}
\description{
This function generates a dataset representing seven branches with added noise.
}
\examples{
seven_branching_data <- seven_branching_data_with_noise(n = 210,
num_noise = 8, min_n = -0.05, max_n = 0.05)
}
