% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{small_big_sphere_with_noise}
\alias{small_big_sphere_with_noise}
\title{Generate data representing small spheres within a larger encompassing sphere with added noise.}
\usage{
small_big_sphere_with_noise(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{Total number of data points to generate, should be a multiple of 13.}

\item{num_noise_dims}{Number of additional noise dimensions to add to the data.}

\item{min_noise}{Minimum value for the noise added to the data.}

\item{max_noise}{Maximum value for the noise added to the data.}
}
\value{
A matrix containing the generated data points with or without added noise.
}
\description{
This function generates data points representing small spheres within a larger encompassing sphere
and adds noise to the data if specified.
}
\examples{
small_big_sphere_with_noise(sample_size = 390, num_noise_dims = 4,
min_noise = -0.05, max_noise = 0.05)

}
