% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd_obj.R
\name{spiral_with_noise}
\alias{spiral_with_noise}
\title{Generate a spiral dataset with optional noise.}
\usage{
spiral_with_noise(n, num_dims, num_noise, min_n, max_n)
}
\arguments{
\item{n}{Total number of data points to generate.}

\item{num_dims}{Number of effective dimensions for each data point.}

\item{num_noise}{Number of additional noise dimensions to add to the data.}

\item{min_n}{Minimum value for the noise added to the data.}

\item{max_n}{Maximum value for the noise added to the data.}
}
\value{
A matrix containing the generated data points with or without added noise.
}
\description{
This function generates a dataset arranged in a spiral pattern with optional noise.
}
\examples{
spiral_with_noise(n = 100, num_dims = 10, num_noise = 4,
min_n = -0.05, max_n = 0.05)

}
