% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear.R
\name{swiss_roll}
\alias{swiss_roll}
\title{Generate Swiss Roll Data}
\usage{
swiss_roll(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{The total number of data points to be generated.}

\item{num_noise_dims}{The number of additional noise dimensions to be generated.}

\item{min_noise}{The minimum value for the noise added to the data points.}

\item{max_noise}{The maximum value for the noise added to the data points.}
}
\value{
A matrix containing the generated Swiss roll data points.
}
\description{
This function generates data points in the shape of a Swiss roll.
}
\examples{

# Generate Swiss roll data with noise with custom parameters
data <- swiss_roll(sample_size = 200, num_noise_dims = 4,
min_noise = -0.05, max_noise = 0.05)
}
