% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circulars.R
\name{three_circulars_with_noise}
\alias{three_circulars_with_noise}
\title{Generate Three Circular Clusters with Noise}
\usage{
three_circulars_with_noise(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{The total number of samples to generate.}

\item{num_noise_dims}{The number of additional noise dimensions to add to the data.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the three circular clusters with added noise.
}
\description{
This function generates three circular clusters in 4D space with added noise dimensions.
}
\examples{
circular_clusters_data <- three_circulars_with_noise(sample_size = 300,
num_noise_dims = 4, min_noise = -0.05, max_noise = 0.05)
}
