% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{three_cluster_mirror_with_noise}
\alias{three_cluster_mirror_with_noise}
\title{Generate Three Cluster Mirror with Noise}
\usage{
three_cluster_mirror_with_noise(n, num_noise, min_n, max_n)
}
\arguments{
\item{n}{The total number of data points to be generated.}

\item{num_noise}{The number of additional noise dimensions to be generated.}

\item{min_n}{The minimum value for the noise added to the data points.}

\item{max_n}{The maximum value for the noise added to the data points.}
}
\value{
A matrix containing the generated data, with each row representing a data point.
}
\description{
This function generates data with three clusters forming a mirror image,
along with added noise.
}
\examples{

# Generate three cluster mirror with noise with custom parameters
data <- three_cluster_mirror_with_noise(n = 300, num_noise = 4,
min_n = -0.05, max_n = 0.05)
}
