% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{three_doublets_with_noise}
\alias{three_doublets_with_noise}
\title{Generate Doublets with Three Clusters and Noise}
\usage{
three_doublets_with_noise(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{The total number of data points to be generated.}

\item{num_noise_dims}{The number of additional noise dimensions to be generated.}

\item{min_noise}{The minimum value for the noise added to the data points.}

\item{max_noise}{The maximum value for the noise added to the data points.}
}
\value{
A matrix containing the generated data, with each row representing a data point.
}
\description{
This function generates data with three sets of doublets (pairs of clusters)
along with added background noise.
}
\examples{

# Generate doublets with three clusters and noise with custom parameters
data <- three_doublets_with_noise(sample_size = 420, num_noise_dims = 4,
min_noise = -0.05, max_noise = 0.05)
}
