% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{three_grid}
\alias{three_grid}
\title{Generate Three Grids with Noise}
\usage{
three_grid(n_value, num_noise, min_n, max_n)
}
\arguments{
\item{n_value}{The number of grid points along the x and y axes for each grid.}

\item{num_noise}{The number of additional noise dimensions to add to the data.}

\item{min_n}{The minimum value for the noise dimensions.}

\item{max_n}{The maximum value for the noise dimensions.}
}
\value{
A list containing three grid datasets with added noise and the sample
 size of each dataset.
}
\description{
This function generates three grid datasets with noise dimensions.
}
\examples{
set.seed(20240412)
three_grids <- three_grid(
  n_value = 19, num_noise = 2, min_n = -0.05,
  max_n = 0.05
)
}
