% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{three_grid_with_noise}
\alias{three_grid_with_noise}
\title{Generate Three Grids with Noise}
\usage{
three_grid_with_noise(n_value, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{n_value}{The number of grid points along the x and y axes for each grid.}

\item{num_noise_dims}{The number of additional noise dimensions to add to the data.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A list containing three grid datasets with added noise and the sample
 size of each dataset.
}
\description{
This function generates three grid datasets with noise dimensions.
}
\examples{
three_grids <- three_grid_with_noise(n_value = 19, num_noise_dims = 4,
min_noise = -0.05, max_noise = 0.05)
}
