% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd_obj.R
\name{torus_3d_row}
\alias{torus_3d_row}
\title{Generate a row of data points for a 3D torus.}
\usage{
torus_3d_row(radius)
}
\arguments{
\item{radius}{A numeric vector containing the radii of the torus, from largest
to smallest.}
}
\value{
A vector representing a row of data points for the 3D torus.
}
\description{
This function generates a row of data points for a 3D torus with given radii.
}
\examples{
set.seed(20240412)
torus_3d_row(c(2, 1))

}
