% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.R
\name{traingular_3d_with_noise}
\alias{traingular_3d_with_noise}
\title{Generate Triangular 3D Datasets with Noise}
\usage{
traingular_3d_with_noise(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{The total number of samples to generate.}

\item{num_noise_dims}{The number of additional noise dimensions to add to the data.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the triangular 3D datasets with added noise.
}
\description{
This function generates triangular 3D datasets with added noise dimensions.
}
\examples{
triangular_3d_data <- traingular_3d_with_noise(sample_size = 150, num_noise_dims = 4,
min_noise = -0.05, max_noise = 0.05)
}
