% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{tree}
\alias{tree}
\title{Generate Tree-like Data with Noise}
\usage{
tree(n, num_noise, min_n, max_n)
}
\arguments{
\item{n}{The total number of samples to generate.}

\item{num_noise}{The number of additional noise dimensions to add to the data.}

\item{min_n}{The minimum value for the noise dimensions.}

\item{max_n}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the tree-like data with added noise.
}
\description{
This function generates a dataset representing a tree-like structure, with added noise.
}
\examples{
set.seed(20240412)
tree_data <- tree(n = 300, num_noise = 2, min_n = -0.05, max_n = 0.05)
}
