% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.R
\name{triangular_plane_with_bkg_noise}
\alias{triangular_plane_with_bkg_noise}
\title{Generate Triangular Plane with Background Noise}
\usage{
triangular_plane_with_bkg_noise(n, num_noise, min_n, max_n)
}
\arguments{
\item{n}{The total number of samples to generate.}

\item{num_noise}{The number of additional noise dimensions to add to the data.}

\item{min_n}{The minimum value for the noise dimensions.}

\item{max_n}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the triangular plane dataset with background noise.
}
\description{
This function generates a triangular plane dataset with background noise dimensions.
}
\examples{
triangular_plane_data <- triangular_plane_with_bkg_noise(n = 675,
num_noise = 3, min_n = -0.05, max_n = 0.05)
}
