% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{two_doublets_parallel_with_noise}
\alias{two_doublets_parallel_with_noise}
\title{Generate Doublets in Parallel with Noise}
\usage{
two_doublets_parallel_with_noise(
  sample_size,
  num_noise_dims,
  min_noise,
  max_noise
)
}
\arguments{
\item{sample_size}{The total number of data points to be generated.}

\item{num_noise_dims}{The number of additional noise dimensions to be generated.}

\item{min_noise}{The minimum value for the noise added to the data points.}

\item{max_noise}{The maximum value for the noise added to the data points.}
}
\value{
A matrix containing the generated data, with each row representing a data point.
}
\description{
This function generates data with two sets of doublets (pairs of clusters)
running in parallel, along with added background noise.
}
\examples{

# Generate doublets in parallel with noise with custom parameters
data <- two_doublets_parallel_with_noise(sample_size = 440, num_noise_dims = 2,
min_noise = -0.05, max_noise = 0.05)
}
