% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{two_grid}
\alias{two_grid}
\title{Generate Two Grids with Noise}
\usage{
two_grid(n_value, num_noise, min_n, max_n)
}
\arguments{
\item{n_value}{The number of grid points along the x and y axes for each grid.}

\item{num_noise}{The number of additional noise dimensions to add to the data.}

\item{min_n}{The minimum value for the noise dimensions.}

\item{max_n}{The maximum value for the noise dimensions.}
}
\value{
A list containing two grid datasets with added noise and the sample
 size of each dataset.
}
\description{
This function generates two grid datasets with noise dimensions.
}
\examples{
set.seed(20240412)
two_grids <- two_grid(n_value = 19, num_noise = 2, min_n = -0.05, max_n = 0.05)
}
