% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{two_grid_comb_bkg}
\alias{two_grid_comb_bkg}
\title{Generate Two Grids with Background Noise}
\usage{
two_grid_comb_bkg(n_value, num_noise, min_n, max_n)
}
\arguments{
\item{n_value}{The number of grid points along each axis for the grids.}

\item{num_noise}{The number of additional noise dimensions to add to the data.}

\item{min_n}{The minimum value for the noise dimensions.}

\item{max_n}{The maximum value for the noise dimensions.}
}
\value{
A list containing the two grid datasets with background noise and the sample size.
}
\description{
This function generates two grid datasets with background noise.
}
\examples{
set.seed(20240412)
two_grid_comb_bkg <- two_grid_comb_bkg(
  n_value = 10, num_noise = 2,
  min_n = -0.05, max_n = 0.05
)
}
