% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{two_nonlinear}
\alias{two_nonlinear}
\title{Generate Two Nonlinear Clusters with Noise}
\usage{
two_nonlinear(n, num_noise, min_n, max_n)
}
\arguments{
\item{n}{The total number of data points to be generated.}

\item{num_noise}{The number of additional noise dimensions to be generated.}

\item{min_n}{The minimum value for the noise added to the data points.}

\item{max_n}{The maximum value for the noise added to the data points.}
}
\value{
A matrix containing the generated data, with each row representing a data point.
}
\description{
This function generates data with two nonlinear clusters along with added noise.
}
\examples{

# Generate two nonlinear clusters with noise with custom parameters
set.seed(20240412)
data <- two_nonlinear(n = 200, num_noise = 2, min_n = -0.05, max_n = 0.50)
}
