% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd_obj.R
\name{cube_3d_with_noise}
\alias{cube_3d_with_noise}
\title{Generate a 3D cube with optional noise.}
\usage{
cube_3d_with_noise(num_of_effective_dims, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{num_of_effective_dims}{Number of effective dimensions (default is 3 for a 3D cube).}

\item{num_noise_dims}{Number of additional noise dimensions to add to the data.}

\item{min_noise}{Minimum value for the noise added to the data.}

\item{max_noise}{Maximum value for the noise added to the data.}
}
\value{
A list containing the generated data matrix and the sample size.
}
\description{
This function generates a 3D cube along with optional noise.
}
\examples{
cube_3d_with_noise(num_of_effective_dims = 3, num_noise_dims = 2,
min_noise = -0.01, max_noise = 0.01)

}
