% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear.R
\name{curvilinear_2d}
\alias{curvilinear_2d}
\title{Generate points on a curvilinear 2D manifold}
\usage{
curvilinear_2d(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{The number of points to generate.}

\item{num_noise_dims}{The number of noise dimensions to add to the generated points.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the generated points on the curvilinear 2D manifold.
}
\description{
This function generates points on a curvilinear 2D manifold based on a nonlinear equation.
}
\examples{
curvilinear_points <- curvilinear_2d(sample_size = 100, num_noise_dims = 2,
                                     min_noise = -1, max_noise = 1)

}
