% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{curvy_branching_cluster_with_bkg_noise}
\alias{curvy_branching_cluster_with_bkg_noise}
\title{Generate Curvy Branching Cluster Data with Background Noise}
\usage{
curvy_branching_cluster_with_bkg_noise(
  sample_size,
  num_noise_dims,
  min_noise,
  max_noise
)
}
\arguments{
\item{sample_size}{The total number of data points to be generated.}

\item{num_noise_dims}{The number of additional noise dimensions to be generated.}

\item{min_noise}{The minimum value for the noise added to the data points.}

\item{max_noise}{The maximum value for the noise added to the data points.}
}
\value{
A matrix containing the generated data, with each row representing a data point.
}
\description{
This function generates data with four clusters, two of which follow a curvilinear pattern and the other two are distributed randomly.
}
\examples{

# Generate curvy branching cluster data with background noise with custom parameters
data <- curvy_branching_cluster_with_bkg_noise(sample_size = 400, num_noise_dims = 10,
min_noise = -0.5, max_noise = 0.5)
}
