% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circulars.R
\name{curvy_cell_cycle_with_noise}
\alias{curvy_cell_cycle_with_noise}
\title{Generate Curvy Cell Cycle Data with Noise}
\usage{
curvy_cell_cycle_with_noise(n, num_noise, min_n, max_n)
}
\arguments{
\item{n}{The total number of samples to generate.}

\item{num_noise}{The number of additional noise dimensions to add to the data.}

\item{min_n}{The minimum value for the noise dimensions.}

\item{max_n}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the curvy cell cycle data with added noise.
}
\description{
This function generates a curvy cell cycle dataset with added noise dimensions.
}
\examples{
curvy_cell_cycle_data <- curvy_cell_cycle_with_noise(n = 300,
num_noise = 4, min_n = -0.05, max_n = 0.05)
}
