% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd_obj.R
\name{dini_surface_with_noise}
\alias{dini_surface_with_noise}
\title{Generate points sampled from the Dini surface with optional noise.}
\usage{
dini_surface_with_noise(n, num_noise, min_n, max_n)
}
\arguments{
\item{n}{Total number of data points to generate.}

\item{num_noise}{Number of additional noise dimensions to add to the data.}

\item{min_n}{Minimum value for the noise added to the data.}

\item{max_n}{Maximum value for the noise added to the data.}
}
\value{
A matrix containing the generated data points with or without added noise.
}
\description{
This function generates points sampled from the Dini surface along with optional noise.
}
\examples{
dini_surface_with_noise(n = 100, num_noise = 4,
min_n = -0.05, max_n = 0.05)

}
