% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gen_bkgnoise}
\alias{gen_bkgnoise}
\title{Generate Background Noise Data}
\usage{
gen_bkgnoise(n = 500, p = 4, m = c(0, 0, 0, 0), s = c(2, 2, 2, 2))
}
\arguments{
\item{n}{Number of samples to generate.}

\item{num_dims}{Number of dimensions (columns) of the data.}

\item{mean}{Mean of the normal distribution used to generate noise (default is 0).}

\item{sd}{Standard deviation of the normal distribution used to generate noise (default is 1).}
}
\value{
A matrix containing the generated background noise data, with
\code{n} rows and \code{num_dims} columns.
}
\description{
This function generates background noise data with specified parameters such as
the number of samples, number of dimensions, mean, and standard deviation.
}
\examples{

# Generate background noise with custom mean and standard deviation
set.seed(20240412)
gen_bkgnoise(n = 500, p = 4, m = c(0, 0, 0, 0), s = c(2, 2, 2, 2))

}
