% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone.R
\name{gen_cone}
\alias{gen_cone}
\title{Generate Blunted Cone}
\usage{
gen_cone(n = 500, p = 4, h = 5, ratio = 0.5)
}
\arguments{
\item{n}{An integer value (default: 500) representing the sample size.}

\item{p}{An integervalue (default: 4) representing the number of dimensions.}

\item{h}{A numeric value (default: 5) representing the height of the cone.}

\item{ratio}{A numeric value (default: 0.5) representing the radius tip to radius base ratio of the cone. Should be less than 1.}
}
\value{
A tibble containing the cone with the option of a sharp or blunted apex.
}
\description{
This function generates a dataset representing a cone with the option of a sharp or blunted apex.
}
\examples{
set.seed(20240412)
cone <- gen_cone(n = 500, p = 4, h = 5, ratio = 0.5)
}
