% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{gen_conicspiral}
\alias{gen_conicspiral}
\title{Generate Conical Spiral}
\usage{
gen_conicspiral(n = 500, spins = 1)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{spins}{A numeric value (default: 1) representing the number of loops of the spiral.}
}
\value{
A data containing a conical spiral structure.
}
\description{
This function generates a dataset representing a conical spiral structure.
}
\examples{
set.seed(20240412)
conicspiral <- gen_conicspiral(n = 500, spins = 1)
}
