% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{gen_crescent}
\alias{gen_crescent}
\title{Generate Crescent}
\usage{
gen_crescent(n = 500, p = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing a Crescent structure.
}
\description{
This function generates a dataset representing a structure with a Crescent pattern.
}
\examples{
set.seed(20240412)
data <- gen_crescent(n = 500, p = 4)
}
