% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{gen_cubic}
\alias{gen_cubic}
\title{Generate Cubic}
\usage{
gen_cubic(n = 500, range = c(-1, 2))
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{range}{A numeric vector (default: c(-1, 2)) representing the range along x1 axis.}
}
\value{
A data containing a cubic structure.
}
\description{
This function generates a dataset representing a structure with a cubic pattern.
}
\examples{
set.seed(20240412)
cubic <- gen_cubic(n = 500)
}
