% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{gen_curvybranches}
\alias{gen_curvybranches}
\title{Generate data with curvy shaped branches}
\usage{
gen_curvybranches(n = 500, p = 4, k = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{k}{A numeric value (default: 4) representing the number of branches.}
}
\value{
A data containing non-linear shaped branches.
}
\description{
This function generates a dataset representing a structure with non-linear shaped branches.
}
\examples{
set.seed(20240412)
data <- gen_curvybranches(n = 500, p = 4, k = 4)
}
