% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{gen_curvycycle}
\alias{gen_curvycycle}
\title{Generate Curvy Cell Cycle in p-d}
\usage{
gen_curvycycle(n = 500, p = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing a curvy cell cycle.
}
\description{
This function generates a dataset representing a structure with a curvy cell cycle.
}
\examples{
set.seed(20240412)
curvycycle <- gen_curvycycle(n = 500, p = 4)
}
