% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{gen_curvyCylinder}
\alias{gen_curvyCylinder}
\title{Generate Curvy Cylinder}
\usage{
gen_curvyCylinder(n = 500, p = 4, r = 1, h = 10, a = 1)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{r}{A numeric value (default: 1) representing the radius of the cylinder.}

\item{h}{A numeric value (default: 10) representing the height of the cylinder.}

\item{a}{A numeric value (default: 1) representing the strength of the cylinder in 4th dimension.}
}
\value{
A data containing a curvy cylinder.
}
\description{
This function generates a dataset representing a structure with a curvy cylinder.
}
\examples{
set.seed(20240412)
data <- gen_curvyCylinder(n = 500, p = 4, r = 1, h = 10, a = 1)
}
