% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{gen_expbranches}
\alias{gen_expbranches}
\title{Generate data with exponential shaped branches}
\usage{
gen_expbranches(n = 400, p = 4, k = 4)
}
\arguments{
\item{n}{A numeric value (default: 400) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{k}{A numeric value (default: 4) representing the number of branches.}
}
\value{
A data containing exponential shaped branches.
}
\description{
This function generates a dataset representing a structure with exponential shaped branches.
}
\examples{
set.seed(20240412)
expbranches <- gen_expbranches(n = 400, p = 4, k = 4)
}
