% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{gen_gaussian}
\alias{gen_gaussian}
\title{Generate Gaussian}
\usage{
gen_gaussian(n = 500, p = 4, s = diag(p) * 0.01)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{s}{A numeric matrix (default:  diag(4) * 0.01) representing the variance of along each dimension.}
}
\value{
A data containing a Gaussian.
}
\description{
This function generates a dataset representing a structure with a Gaussian.
}
\examples{
set.seed(20240412)
gaussian_data <- gen_gaussian(n = 500, p = 4, s = diag(4))
}
