% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{gen_gridcube}
\alias{gen_gridcube}
\title{Generate Cube with grid points}
\usage{
gen_gridcube(n = 500, p = 4)
}
\arguments{
\item{n}{A numeric vector (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the cube with grid points.
}
\description{
This function generates a grid dataset with specified grid points along each axes.
}
\examples{
set.seed(20240412)
gridcube <- gen_gridcube(n = 500, p = 4)
}
