% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{gen_gridedSphere}
\alias{gen_gridedSphere}
\title{Generate Grided Sphere}
\usage{
gen_gridedSphere(n = 20, p = 4, r = 1)
}
\arguments{
\item{n}{A numeric value (default: 20) the number of points or the angular step size between sampled coordinates.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{r}{A numeric value (default: 1) representing the radius of the sphere.}
}
\value{
A data containing a grided sphere.
}
\description{
This function generates a dataset representing a structure with a grided sphere.
}
\examples{
set.seed(20240412)
sphere_data <- gen_gridedSphere(n = 20, p = 4, r = 1)
head(sphere_data, 5)
}
