% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{gen_hemisphere}
\alias{gen_hemisphere}
\title{Generate Hemisphere}
\usage{
gen_hemisphere(n = 500)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}
}
\value{
A data containing a hemisphere.
}
\description{
This function generates a dataset representing a structure with a hemisphere.
}
\examples{
set.seed(20240412)
hemisphere <- gen_hemisphere(n = 500)
}
