% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{gen_linearbranches}
\alias{gen_linearbranches}
\title{Generate data with linear shaped branches}
\usage{
gen_linearbranches(n = 400, k = 4)
}
\arguments{
\item{n}{A numeric value (default: 400) representing the sample size.}

\item{k}{A numeric value (default: 4) representing the number of branches.}
}
\value{
A data containing linear shaped branches.
}
\description{
This function generates a dataset representing a structure with linear shaped branches.
}
\examples{
set.seed(20240412)
linearbranches <- gen_linearbranches(n = 400, k = 4)
}
