% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{gen_longLinear}
\alias{gen_longLinear}
\title{Generate Long Linear Data}
\usage{
gen_longLinear(n = 500, p = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the long linear data.
}
\description{
This function generates a dataset consisting of long linear data.
}
\examples{
set.seed(20240412)
long_cluster <- gen_longLinear(n = 500, p = 4)
}
