% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobius.R
\name{gen_mobius}
\alias{gen_mobius}
\title{Generate a 3-D Mobius in High Dimensions}
\usage{
gen_mobius(n = 500, p = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing a mobius structure.
}
\description{
This function generates a dataset representing a structure with a mobius.
}
\examples{
set.seed(20240412)
curvy_data <- gen_mobius(n = 500, p = 4)
}
