% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_shapes.R
\name{gen_mobiusgau}
\alias{gen_mobiusgau}
\title{Generate Gaussian cluster with the Mobius Cluster}
\usage{
gen_mobiusgau(n = c(200, 100), p = 4)
}
\arguments{
\item{n}{A numeric vector (default: c(200, 100)) representing the sample sizes.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}
}
\value{
A data containing the mobius cluster and Gaussian cluster.
}
\description{
This function generates a dataset consisting of a mobius cluster and Gaussian cluster.
}
\examples{
data <- gen_mobiusgau(n = c(200, 100), p = 4)
}
