% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{gen_orgcurvybranches}
\alias{gen_orgcurvybranches}
\title{Generate data with curvy shaped branches in a initial point}
\usage{
gen_orgcurvybranches(n = 400, p = 4, k = 4)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{k}{A numeric value (default: 4) representing the number of branches.}
}
\value{
A data containing curvy shaped branches originated in one point.
}
\description{
This function generates a dataset representing a structure with curvy shaped branches.
}
\examples{
set.seed(20240412)
data <- gen_orgcurvybranches(n = 400, p = 4, k = 4)
}
