% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{gen_pyrstar}
\alias{gen_pyrstar}
\title{Generate Star Based Pyramid}
\usage{
gen_pyrstar(n = 500, p = 4, h = 5, rb = 3)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{h}{A numeric value (default: 5) representing the h of the pyramid.}

\item{rb}{A numeric value (default: 3) representing the base radius of the pyramid.}
}
\value{
A data containing the star based pyramid.
}
\description{
This function generates a dataset representing a star based pyramid.
}
\examples{
set.seed(20240412)
hexagonal_corn_data <- gen_pyrstar(n = 500, p = 4, h = 5, rb = 3)
}
