% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{gen_pyrtri}
\alias{gen_pyrtri}
\title{Generate Triangular Based Pyramid}
\usage{
gen_pyrtri(n = 500, p = 4, h = 5, l = 3, rt = 0.5)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{h}{A numeric value (default: 5) representing the h of the pyramid.}

\item{l}{A numeric value (default: 3) representing the base length of the pyramid.}

\item{rt}{A numeric value (default: 0.5) representing the tip radius of the pyramid.}
}
\value{
A data containing the triangular based pyramid.
}
\description{
This function generates a dataset representing a triangular based pyramid.
}
\examples{
set.seed(20240412)
triangular_corn_data <- gen_pyrtri(n = 500, p = 4, h = 5, l = 3, rt = 0.5)
}
