% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{gen_sphericalspiral}
\alias{gen_sphericalspiral}
\title{Generate Spherical Spiral}
\usage{
gen_sphericalspiral(n = 500, p = 4, spins = 1)
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{spins}{A numeric value (default: 1) representing the number of loops of the spiral.}
}
\value{
A data containing a spherical spiral.
}
\description{
This function generates a dataset representing a structure with a spherical spiral.
}
\examples{
set.seed(20240412)
sphericalspiral <- gen_sphericalspiral(n = 500, p = 4, spins = 1)
}
