% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss_roll.R
\name{gen_swissroll}
\alias{gen_swissroll}
\title{Generate Swiss Roll Data}
\usage{
gen_swissroll(n = 500, p = 4, w = c(-1, 1))
}
\arguments{
\item{n}{A numeric value (default: 500) representing the sample size.}

\item{p}{A numeric value (default: 4) representing the number of dimensions.}

\item{w}{A numeric vector (default: c(-1, 1)) representing the vertical variation.}
}
\value{
A data containing the generated swiss roll data.
}
\description{
This function generates swiss roll data.
}
\examples{
set.seed(20240412)
data <- gen_swissroll(n = 500, p = 4)
head(data, 5)
}
\references{
Agrafiotis, D. K., & Xu, H. (2002).
A self-organizing principle for learning nonlinear manifolds.
\emph{Proceedings of the National Academy of Sciences}, \emph{99}(25), 15869-15872.


Roweis, S. T., & Saul, L. K. (2000).
Nonlinear dimensionality reduction by locally linear embedding.
\emph{Science}, \emph{290}(5500), 2323-2326.
}
