% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_curve.R
\name{mirror_s_curves_with_noise}
\alias{mirror_s_curves_with_noise}
\title{Generate Mirror S-curve Datasets with Noise}
\usage{
mirror_s_curves_with_noise(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{The total number of samples to generate (should be divisible by 2).}

\item{num_noise_dims}{The number of additional noise dimensions to add to the data.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the combined mirror S-curve datasets with added noise.
}
\description{
This function generates mirror S-curve datasets with added noise dimensions.
}
\examples{
mirror_s_curve_data <- mirror_s_curves_with_noise(sample_size = 200, num_noise_dims = 8,
min_noise = -0.5, max_noise = 0.5)
}
