% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{normalize_data}
\alias{normalize_data}
\title{Generate Normalized data}
\usage{
normalize_data(data)
}
\arguments{
\item{data}{A matrix representing the data which needed to be normalized.}
}
\value{
A normalized data.
}
\description{
This function to normalize data by absolute value
}
\examples{
set.seed(20240412)
data1 <- gen_gaussian(n= 500, p = 4)
scale_data(data = data1)

}
