% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{one_grid_diff_with_bkg_noise}
\alias{one_grid_diff_with_bkg_noise}
\title{Generate One Grid with Different Values and Background Noise}
\usage{
one_grid_diff_with_bkg_noise(
  sample_size = 260,
  num_noise_dims = 5,
  min_noise = -0.05,
  max_noise = 0.05
)
}
\arguments{
\item{sample_size}{The total number of samples, including the background noise.}

\item{num_noise_dims}{The number of additional noise dimensions to add to the data.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A list containing the grid dataset with different values and background noise.
}
\description{
This function generates a grid dataset with different values and background noise.
}
\examples{
one_grid_diff_with_bkg_noise <- one_grid_diff_with_bkg_noise(sample_size = 260,
num_noise_dims = 5, min_noise = -0.05, max_noise = 0.05)
}
