% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear.R
\name{sine_curve_with_noise}
\alias{sine_curve_with_noise}
\title{Generate Sine Curve Data with Noise}
\usage{
sine_curve_with_noise(sample_size, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{The number of samples to generate.}

\item{num_noise_dims}{The number of additional noise dimensions to add to the data.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the sine curve data with noise.
}
\description{
This function generates a dataset representing a sine curve with added noise.
}
\examples{
sine_curve_with_noise <- sine_curve_with_noise(sample_size = 100, num_noise_dims = 8,
                                               min_noise = -0.05, max_noise = 0.05)
}
