% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd_obj.R
\name{spiral_with_noise}
\alias{spiral_with_noise}
\title{Generate a spiral dataset with optional noise.}
\usage{
spiral_with_noise(sample_size, num_dims, num_noise_dims, min_noise, max_noise)
}
\arguments{
\item{sample_size}{Total number of data points to generate.}

\item{num_dims}{Number of effective dimensions for each data point.}

\item{num_noise_dims}{Number of additional noise dimensions to add to the data.}

\item{min_noise}{Minimum value for the noise added to the data.}

\item{max_noise}{Maximum value for the noise added to the data.}
}
\value{
A matrix containing the generated data points with or without added noise.
}
\description{
This function generates a dataset arranged in a spiral pattern with optional noise.
}
\examples{
spiral_with_noise(sample_size = 100, num_dims = 10, num_noise_dims = 4,
min_noise = -0.05, max_noise = 0.05)

}
