% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{three_clust_01}
\alias{three_clust_01}
\title{Three-Cluster Dataset in 4-D}
\format{
A data frame with 1500 rows and 5 columns:
\describe{
  \item{cluster, x1, x2, x3, x4}{High-dimensional coordinates}
}
}
\source{
This dataset is generated for example purposes.
}
\usage{
data(three_clust_01)
}
\description{
The `three_clust_01` dataset contains three distinct clusters in a 4-D space.
}
\examples{
# Load the mobiusgau dataset
data(three_clust_01)

# Display the first few rows of the dataset
head(three_clust_01)

}
\keyword{datasets}
