% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.R
\name{traingular_3d_with_noise}
\alias{traingular_3d_with_noise}
\title{Generate Triangular 3D Datasets with Noise}
\usage{
traingular_3d_with_noise(n, num_noise, min_n, max_n)
}
\arguments{
\item{n}{The total number of samples to generate.}

\item{num_noise}{The number of additional noise dimensions to add to the data.}

\item{min_n}{The minimum value for the noise dimensions.}

\item{max_n}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the triangular 3D datasets with added noise.
}
\description{
This function generates triangular 3D datasets with added noise dimensions.
}
\examples{
triangular_3d_data <- traingular_3d_with_noise(n = 150, num_noise = 4,
min_n = -0.05, max_n = 0.05)
}
