% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.R
\name{triangular_plane_with_bkg_noise}
\alias{triangular_plane_with_bkg_noise}
\title{Generate Triangular Plane with Background Noise}
\usage{
triangular_plane_with_bkg_noise(
  sample_size,
  num_noise_dims,
  min_noise,
  max_noise
)
}
\arguments{
\item{sample_size}{The total number of samples to generate.}

\item{num_noise_dims}{The number of additional noise dimensions to add to the data.}

\item{min_noise}{The minimum value for the noise dimensions.}

\item{max_noise}{The maximum value for the noise dimensions.}
}
\value{
A matrix containing the triangular plane dataset with background noise.
}
\description{
This function generates a triangular plane dataset with background noise dimensions.
}
\examples{
triangular_plane_data <- triangular_plane_with_bkg_noise(sample_size = 675,
num_noise_dims = 3, min_noise = -0.05, max_noise = 0.05)
}
