% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_curve.R
\name{two_s_curve_hole_with_noise}
\alias{two_s_curve_hole_with_noise}
\title{Generate Two S-Curve Data with Noise}
\usage{
two_s_curve_hole_with_noise(
  n = 200,
  num_noise = 8,
  min_n = -0.05,
  max_n = 0.05
)
}
\arguments{
\item{n}{The total number of data points to be generated.}

\item{num_noise}{The number of additional noise dimensions to be generated.}

\item{min_n}{The minimum value for the noise added to the data points.}

\item{max_n}{The maximum value for the noise added to the data points.}
}
\value{
A list containing the generated data matrix (`df`) and the total sample size.
}
\description{
This function generates two S-curve data with noise.
}
\examples{

# Generate two S-curve data with noise with custom parameters
data <- two_s_curve_hole_with_noise(n = 200, num_noise = 8,
min_n = -0.05, max_n = 0.05)
}
